pyproject.toml
[project]
name = "log-csi"
version = "0.1.0"
description = "Cross-platform CLI to investigate logs and reconstruct an incident timeline."
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.9.0",
    "pyyaml>=6.0.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "rich>=13.0.0",
    "tzdata",
    "python-dotenv>=1.0.0"
]

[project.scripts]
log-csi = "log_csi.cli:app"
